pyproject.toml
[project]
name = "code-buddy"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
